server/requirements.txt
asgiref==3.11.0
boto3==1.42.73
botocore==1.42.73
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==4.2
django-cors-headers==4.9.0
# django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
gunicorn==25.1.0
idna==3.11
jmespath==1.1.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyJWT==2.11.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
# redis==7.4.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0


# after fix an erro of aws ec2 verification